+++ src/components/Authentication/Logout.js
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import { Redirect } from 'react-router-dom'
-import errorMessage from '../errorMessage';
-import { auth } from '../firebase';
+import errorMessage from './../../errorMessage';
+import { auth } from '../../firebase';
 
 export default function Logout() {
     const [loading, setLoading] = useState(true);
