+++ src/components/ChallengePage/ChallengePage.js
@@ -1,5 +1,5 @@
 import React, {useState} from 'react'
-import { useAuth } from '../contexts/AuthContext';
+import { useAuth } from './../../contexts/AuthContext';
 import ChallengePagePopup from './ChallengePagePopup';
 import ChallengeIcon from "./ChallengeIcon";
 import { Button, Col, Container, Row } from 'react-bootstrap';
