+++ src/components/Helpers/URLEncode.js
