+++ src/components/HomePage/HomePageInfo.scss
